+++ order/src/main/java/com/owlexpress/order/presentation/dto/request/UpdateOrderStatusRequestDto.java
@@ -1,6 +1,6 @@
 package com.owlexpress.order.presentation.dto.request;
 
-import com.owlexpress.order.domain.constant.OrderStatus;
+import com.owlexpress.order.common.constant.OrderStatus;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
