+++ order/src/main/java/com/owlexpress/order/presentation/dto/response/OrderSearchResponseDto.java
@@ -1,7 +1,7 @@
 package com.owlexpress.order.presentation.dto.response;
 
-import com.owlexpress.order.domain.constant.OrderStatus;
-import com.owlexpress.order.domain.constant.OrderType;
+import com.owlexpress.order.common.constant.OrderStatus;
+import com.owlexpress.order.common.constant.OrderType;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
